simplify gamification models
