+++ libs/common/src/models/gamification-domain.model.ts
@@ -47,26 +47,56 @@ export interface ActivityXpResult {
   breakdown: XpBreakdownItem[];
 }
 
+export interface BadgeReward {
+  id: string;
+  name: string;
+  description: string;
+  achievedAt: string;
+}
+
+export interface AchievementReward {
+  id: string;
+  name: string;
+  description: string;
+  progress: number;
+  completed: boolean;
+  completedAt?: string;
+}
+
+export interface XpReward {
+  processed: boolean;
+  awarded: number;
+  breakdown: XpBreakdownItem[];
+}
+
+export interface ActivityProcessingResult {
+  processed: boolean;
+  processedAt: string;
+  xp?: XpReward;
+  badges?: BadgeReward[];
+  achievements?: AchievementReward[];
+}
+
 export const DEFAULT_XP_SETTINGS: GameXpSettings = {
-  'CODE_PUSH': {
+  CODE_PUSH: {
     base: 10,
     bonuses: {
-      multipleCommits: 5
-    }
+      multipleCommits: 5,
+    },
   },
-  'PR_CREATED': {
-    base: 20
+  PR_CREATED: {
+    base: 20,
   },
-  'PR_MERGED': {
+  PR_MERGED: {
     base: 30,
     bonuses: {
-      merged: 20
-    }
+      merged: 20,
+    },
   },
-  'ISSUE_CREATED': {
-    base: 15
+  ISSUE_CREATED: {
+    base: 15,
   },
-  'ISSUE_CLOSED': {
-    base: 20
-  }
+  ISSUE_CLOSED: {
+    base: 20,
+  },
 };
