+++ libs/common/src/models/index.ts
@@ -2,5 +2,5 @@ export * from './common.model';
 export * from './event.model';
 export * from './github-shared.model';
 export * from './user.model';
-export * from './xp-domain.model';
+export * from './gamification-domain.model';
